make utils a module
